requirements.txt
google-auth-oauthlib==0.4.6
google-api-python-client==2.47.0
requests==2.28.1
requests-oauthlib==1.3.1
oauthlib==3.2.0
tqdm==4.64.0
python-dotenv==0.20.0
